+++ cli/src/complete.rs
@@ -929,8 +929,8 @@ fn get_jj_command() -> Result<(JjBuilder, UserSettings), CommandError> {
     let maybe_cwd_workspace_loader = DefaultWorkspaceLoaderFactory.create(find_workspace_dir(&cwd));
     let _ = config_env.reload_user_config(&mut raw_config);
     if let Ok(loader) = &maybe_cwd_workspace_loader {
-        config_env.reset_repo_path(loader.repo_path());
-        let _ = config_env.reload_repo_config(&mut raw_config);
+        config_env.reset_repo_path(loader.repo_path(), loader.workspace_root());
+        let _ = config_env.reload_all_repo_config(&ui, &mut raw_config);
     }
     let mut config = config_env.resolve_config(&raw_config)?;
     // skip 2 because of the clap_complete prelude: jj -- jj <actual args...>
@@ -947,8 +947,8 @@ fn get_jj_command() -> Result<(JjBuilder, UserSettings), CommandError> {
     if let Some(repository) = args.repository {
         // Try to update repo-specific config on a best-effort basis.
         if let Ok(loader) = DefaultWorkspaceLoaderFactory.create(&cwd.join(&repository)) {
-            config_env.reset_repo_path(loader.repo_path());
-            let _ = config_env.reload_repo_config(&mut raw_config);
+            config_env.reset_repo_path(loader.repo_path(), loader.workspace_root());
+            let _ = config_env.reload_all_repo_config(&ui, &mut raw_config);
             if let Ok(new_config) = config_env.resolve_config(&raw_config) {
                 config = new_config;
             }