+++ cli/src/merge_tools/mod.rs
@@ -46,6 +46,7 @@ use thiserror::Error;
 
 use self::builtin::edit_diff_builtin;
 use self::builtin::edit_merge_builtin;
+pub use self::builtin::make_diff_sections;
 use self::builtin::BuiltinToolError;
 pub(crate) use self::diff_working_copies::new_utf8_temp_dir;
 use self::diff_working_copies::DiffCheckoutError;