+++ cli/src/revset_util.rs
@@ -145,6 +145,7 @@ fn warn_user_redefined_builtin(
         ConfigSource::Default => (),
         ConfigSource::EnvBase
         | ConfigSource::User
+        | ConfigSource::RepoManaged
         | ConfigSource::Repo
         | ConfigSource::EnvOverrides
         | ConfigSource::CommandArg => {