+++ cli/tests/runner.rs
@@ -58,6 +58,7 @@ mod test_operations;
 mod test_parallelize_command;
 mod test_rebase_command;
 mod test_repo_change_report;
+mod test_repo_managed_config;
 mod test_resolve_command;
 mod test_restore_command;
 mod test_revert_command;